Updating SVG Coordinate System
In original code, we usually find this line:
`y = -y + ... `
.. which making sense for usual coordinate system(x goes right, y goes up), but make huge issue in SVG (x goes right, but y goes down).
One of problem arise when we use transformation (translation and rotation) attribute. When we place an element, it follows usual coordinate system. But when translating/rotating, suddenly it follows SVG coordinate system.

Solution:
I delete anything related to mentioned line of code, so the whole code follows SVG coordinate system.